+++ modules/angular2/angular2.ts
@@ -1,6 +1,6 @@
 export * from './common';
 export * from './core';
-export * from './profile';
+export * from './instrumentation';
 export * from './platform/browser';
 export * from './src/platform/dom/dom_adapter';
 export * from './upgrade';
