+++ modules/angular2/instrumentation.ts
