+++ modules/angular2/src/http/backends/xhr_backend.ts
@@ -4,7 +4,7 @@ import {Request} from '../static_request';
 import {Response} from '../static_response';
 import {Headers} from '../headers';
 import {ResponseOptions, BaseResponseOptions} from '../base_response_options';
-import {Injectable} from 'angular2/angular2';
+import {Injectable} from 'angular2/core';
 import {BrowserXhr} from './browser_xhr';
 import {isPresent} from 'angular2/src/facade/lang';
 import {Observable} from 'angular2/angular2';
