+++ modules/angular2/src/platform/browser/tools/common_tools.ts
