+++ modules/angular2/src/platform/browser/tools/tools.ts
