+++ modules/angular2/src/platform/dom/debug/debug_element_view_listener.ts
