+++ modules/angular2/test/common/directives/ng_switch_spec.ts
@@ -12,7 +12,7 @@ import {
   xit,
 } from 'angular2/testing_internal';
 
-import {Component, View} from 'angular2/angular2';
+import {Component, View} from 'angular2/core';
 
 import {NgSwitch, NgSwitchWhen, NgSwitchDefault} from 'angular2/src/common/directives/ng_switch';
