+++ modules/angular2/test/platform/browser/bootstrap_spec.ts
