+++ modules/angular2/test/platform/browser/tools/spies.ts
@@ -1,5 +1,5 @@
 import {SpyObject} from 'angular2/testing_internal';
-import {Injector, provide} from 'angular2/angular2';
+import {Injector, provide} from 'angular2/core';
 import {ComponentRef} from 'angular2/src/core/linker/dynamic_component_loader';
 import {global} from 'angular2/src/facade/lang';
 import {ApplicationRef, ApplicationRef_} from 'angular2/src/core/application_ref';
