+++ modules/angular2/test/platform/browser/xhr_impl_spec.ts
