+++ modules/angular2/test/public_api_spec.ts
@@ -596,17 +596,16 @@ var NG_ALL = [
   'DirectiveResolver',
   'DirectiveResolver.resolve()',
   'DynamicComponentLoader',
-  'DomAdapter',
-  'DomAdapter.attrToPropMap:dart',
-  'DomAdapter.attrToPropMap=:dart',
-  'setRootDomAdapter()',
-  'DOM',
+  'DomAdapter:js',
+  'setRootDomAdapter():js',
+  'DOM:js',
   /*
    Abstract methods
   'DynamicComponentLoader.loadAsRoot()',
   'DynamicComponentLoader.loadIntoLocation()',
   'DynamicComponentLoader.loadNextToLocation()',
   */
+  'enableDebugTools():js',
   'ELEMENT_PROBE_PROVIDERS:js',
   'ELEMENT_PROBE_BINDINGS:js',
   'ElementRef',
@@ -1412,6 +1411,7 @@ var NG_ALL = [
   'AfterViewInit:dart',
   'ControlValueAccessor:dart',
   'DoCheck:dart',
+  'disableDebugTools():js',
   'Form:dart',
   'HostViewRef:dart',
   'HostViewRef.changeDetectorRef',
