+++ modules/angular2/test/router/integration/router_integration_spec.ts
@@ -14,7 +14,7 @@ import {
   xit,
 } from 'angular2/testing_internal';
 
-import {bootstrap} from 'angular2/bootstrap';
+import {bootstrap} from 'angular2/platform/browser';
 import {Component, Directive, View} from 'angular2/src/core/metadata';
 import {DOM} from 'angular2/src/platform/dom/dom_adapter';
 import {provide, ViewChild, AfterViewInit} from 'angular2/core';
