+++ modules/angular2/test/router/router_link_spec.ts
@@ -17,7 +17,7 @@ import {
 import {SpyRouter, SpyLocation} from './spies';
 
 import {provide, Component, View} from 'angular2/core';
-import {By} from 'angular2/platform/browser';
+import {By} from 'angular2/platform/common_dom';
 
 import {
   Location,
