+++ modules/angular2/test/testing/testing_public_spec.ts
@@ -16,7 +16,7 @@ import {
 
 import {Injectable, bind} from 'angular2/core';
 import {NgIf} from 'angular2/common';
-import {Directive, Component, View, ViewMetadata} from 'angular2/angular2';
+import {Directive, Component, View, ViewMetadata} from 'angular2/core';
 import {XHR} from 'angular2/src/compiler/xhr';
 import {XHRImpl} from 'angular2/src/platform/browser/xhr_impl';
