--- modules/angular2/tools.ts
@@ -1,4 +1,0 @@
-/*
- * Debugging and profiling tools for Angular 2
- */
-export {enableDebugTools, disableDebugTools} from 'angular2/src/tools/tools';
